code cleanup and icon support for course switch sidebar
